// Neptune.Web/src/app/shared/generated/model/treatment-bmp-delineation-map-dto.ts
export class TreatmentBMPDelineationMapDto { 
    TreatmentBMPID?: number;
    TreatmentBMPName?: string | null;
    TreatmentBMPTypeName?: string | null;
    StormwaterJurisdictionID?: number;
    Latitude?: number;
    Longitude?: number;
    HasDelineation?: boolean;
    DelineationID?: number | null;
    DelineationTypeID?: number | null;
    IsVerified?: boolean | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}